Nettoyage
Nettoyage + transfert des fonctions annexes dans une librairie dédiée
